infrastructure/pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".

[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools>=46.1.0", "wheel", "build", "docutils" ]

[project]
name = "infrastructure"
dynamic = ["version"]
classifiers = [
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11"
]
dependencies = [
  "aws-cdk-lib>=2.40.0",
  "constructs>=10.0.0,<11.0.0",
  "boto3",
  "pydantic",
  "rich",
  "awscdk-minecraft",
  "aws-cdk.aws-batch-alpha==2.40.0a0",
]
description = "Infrastructure as code."
readme = "README.md"
requires-python = ">=3.7"
authors = [ ]

  [project.license]
  text = "Proprietary"

  [project.optional-dependencies]
  test = [ "pytest-cov", "pytest-xdist", "pytest" ]
  all = [ "infrastructure[test]" ]
  dev = [ "infrastructure[test]" ]

[tool.setuptools.dynamic]
version = { file = "VERSION.txt" }

[tool.autoflake]
ignore_init_module_imports = true
remove_all_unused_imports = false
remove_unused_variable = [ "all" ]

[tool.black]
line-length = 120

[tool.flake8]
docstring-convention = "all"
ignore = [ "D107", "D212", "E501", "W503", "W605", "D203", "D104" ]
radon-max-cc = 10

[tool.isort]
profile = "black"

[tool.pylint."messages control"]
disable = [
  "line-too-long",
  "trailing-whitespace",
  "missing-function-docstring",
  "consider-using-f-string",
  "import-error",
  "too-few-public-methods",
  "redefined-outer-name"
]
